Extract more filesystem usage
